+++ src/shared/APIResponse.ts
@@ -1,4 +1,4 @@
-import { HttpError } from './HttpErrors'
+import { HttpError } from './HttpError'
 import { Result } from './Result'
 
 export type ApiResult<T> = Result<T, HttpError>
