+++ src/shared/HttpError.ts
